RPM+1

+++ gentelella/study/summary_plots.py
@@ -296,7 +296,8 @@ def makeDEbox(input_file):
                     y_dict[cond] = row[1:]
         data = []
         for i,key in enumerate(x_dict.keys()):
-            to_y = numpy.ndarray(list(map(int, y_dict[key])))
+
+            to_y = numpy.ndarray(list(map(float, y_dict[key])))
             to_y.astype(float)
 
             to_y = to_y+1
